+++ src/ideApi/_IdeApi.ts
@@ -30,6 +30,7 @@ export {
   LanguageClientOptions,
   ResponseError,
   ServerOptions as ClientServerOptions,
+  TextDocumentIdentifier,
 } from "vscode-languageclient";
 
 export { URI } from "vscode-uri";