+++ src/localExecution/dafnyRunner.ts
@@ -1,6 +1,4 @@
 "use strict";
-import * as os from "os";
-
 import { WorkspaceConfiguration, workspace, window } from "../ideApi/_IdeApi";
 import {
   Config,
@@ -12,7 +10,8 @@ import { IDafnyRunner } from "./IDafnyRunner";
 
 /**
  * This class is used for running Dafny files after they were compiled.
- * Therefore it also supports mono for macOS / Linux.
+ * Since it uses dotnet in-place of the self-contained executables, no platform
+ * specific handling is necessary.
  */
 export class DafnyRunner implements IDafnyRunner {
   private config: WorkspaceConfiguration = workspace.getConfiguration(
@@ -32,15 +31,9 @@ export class DafnyRunner implements IDafnyRunner {
       EnvironmentConfig.DafnySuffix,
       EnvironmentConfig.ExeSuffix
     );
-    const useMono: boolean =
-      this.config.get<boolean>(Config.UseMono) ||
-      os.platform() !== EnvironmentConfig.Win32;
-    if (useMono) {
-      const monoExecutable =
-        this.config.get<string>(Config.MonoExecutablePath) ||
-        EnvironmentConfig.Mono;
-      return `${monoExecutable} "${executable}"`;
-    }
-    return `& "${executable}"`;
+    const dotnetExecutable =
+      this.config.get<string>(Config.DotnetExecutablePath) ||
+      EnvironmentConfig.Dotnet;
+    return `& ${dotnetExecutable} "${executable}"`;
   }
 }