+++ src/typeInterfaces/ICounterExampleArguments.ts
@@ -1,4 +1,7 @@
 "use strict";
+
+import { TextDocumentIdentifier } from "../ideApi/_IdeApi";
+
 export interface ICounterExampleArguments {
-  DafnyFile: string;
+  TextDocument: TextDocumentIdentifier;
 }